backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==25.1.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
bitarray==3.0.0
blake2b-py==0.3.2
blinker==1.9.0
cached-property==1.5.2
cachetools==5.5.1
cairo-lang==0.13.3
cbor2==5.6.5
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
chardet==4.0.0
charset-normalizer==3.4.1
ckzg==1.0.2
click==8.1.8
cloud-init==24.4.1
cmake==3.31.6
coincurve==20.0.0
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
configobj==5.0.6
configparser==7.1.0
constantly==15.1.0
contourpy==1.3.1
coverage==6.5.0
cryptography==3.4.8
crytic-compile==0.3.8
cycler==0.12.1
cytoolz==0.12.3
dbus-python==1.2.18
dictionaries==0.0.2
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.7.0
ecdsa==0.19.0
eth-abi==4.2.1
eth-account==0.11.3
eth-bloom==1.0.4
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.2
eth-utils==2.3.2
ethereum==2.3.2
ethereum-abi-utils==0.4.7
ethereum-tester==0.1.0b3
ethereum-utils==0.6.2
ethereum_input_decoder==0.2.2
exceptiongroup==1.2.2
execnet==2.1.1
fastapi==0.115.8
fastecdsa==3.0.1
filelock==3.17.0
Flask==3.1.0
Flask-Cors==3.0.10
fonttools==4.56.0
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.3.0
future==1.0.0
gprof2dot==2024.6.6
greenlet==3.1.1
gunicorn==20.1.0
h11==0.14.0
hexbytes==0.2.3
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
hyperlink==21.0.0
hypothesis==6.125.2
identify==2.6.7
idna==3.10
importlib-metadata==4.6.4
importlib_resources==6.5.2
incremental==21.3.0
iniconfig==2.0.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==23.5.0
kiwisolver==1.4.8
langchain==0.3.18
langchain-core==0.3.35
langchain-text-splitters==0.3.6
langsmith==0.3.8
lark==1.2.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lit==18.1.8
lru-dict==1.2.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.1.1
marshmallow_dataclass==8.6.1
matplotlib==3.10.0
mock==5.1.0
more-itertools==8.10.0
motor==3.7.0
mpmath==1.3.0
multidict==6.1.0
mypy_extensions==0.4.4
mythril==0.24.8
netifaces==0.11.0
networkx==3.4.2
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
openai==1.62.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parsimonious==0.9.0
pbkdf2==1.3
persistent==6.1
pillow==11.1.0
pipdeptree==2.25.0
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==2.20.0
prettytable==3.14.0
prometheus_client==0.21.1
propcache==0.2.1
protobuf==5.29.3
py-ecc==1.4.7
py-evm==0.7.0a1
py-flags==1.1.4
py-solc==3.2.0
py-solc-x==1.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.22
pycryptodome==3.21.0
pycurl==7.44.1
pydantic==2.10.6
pydantic_core==2.27.2
pyethash==0.1.27
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pylru==1.2.1
pymongo==4.11.3
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
pysha3==1.0.2
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-cov==5.0.0
pytest-mock==3.14.0
pytest-profiling==1.8.1
pytest-xdist==3.6.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
pyunormalize==16.0.0
PyYAML==5.4.1
referencing==0.36.2
regex==2024.11.6
repoze.lru==0.7
requests==2.32.3
requests-toolbelt==1.0.0
rlp==3.0.0
rpds-py==0.22.3
scikit-learn==1.6.1
scipy==1.15.1
scrypt==0.8.27
SecretStorage==3.3.1
semantic-version==2.10.0
service-identity==18.1.0
six==1.16.0
slither==0.3.0
slither-analyzer==0.11.0
sniffio==1.3.1
solc-select==1.0.4
sortedcontainers==2.4.0
SQLAlchemy==2.0.38
starlette==0.45.3
sympy==1.13.3
systemd-python==234
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
torch==2.0.0
tqdm==4.67.1
transaction==5.0
transformers==4.28.1
trie==2.2.0
triton==2.0.0
Twisted==22.1.0
typeguard==2.13.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.2
wadllib==1.3.6
wcwidth==0.2.13
web3==6.20.3
websockets==14.2
Werkzeug==3.1.3
yarl==1.18.3
z3-solver==4.12.5.0
zipp==1.0.0
zope.deferredimport==5.0
zope.interface==7.2
zope.proxy==6.1
zstandard==0.23.0
